Change Read() size and return value to a size_t for clarity
BUG=100641



TEST=N/A
Review URL: http://codereview.chromium.org/8352019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@106533 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/filters/ffmpeg_glue_unittest.cc
@@ -23,7 +23,7 @@ class MockProtocol : public FFmpegURLProtocol {
   MockProtocol() {
   }
 
-  MOCK_METHOD2(Read, int(int size, uint8* data));
+  MOCK_METHOD2(Read, size_t(size_t size, uint8* data));
   MOCK_METHOD1(GetPosition, bool(int64* position_out));
   MOCK_METHOD1(SetPosition, bool(int64 position));
   MOCK_METHOD1(GetSize, bool(int64* size_out));
